Separated ADC parameter group code into separate files.
