+++ src/main/drivers/adc.h
@@ -19,6 +19,8 @@
 
 #include "drivers/io_types.h"
 
+#include "parameter_groups/adc.h"
+
 #ifndef ADC_INSTANCE
 #define ADC_INSTANCE                ADC1
 #endif
@@ -68,19 +70,6 @@ typedef struct adc_config_s {
     bool enabled;
     uint8_t sampleTime;
 } adcOperatingConfig_t;
-
-typedef struct adcChannelConfig_t {
-    bool enabled;
-    ioTag_t ioTag;
-} adcChannelConfig_t;
-
-typedef struct adcConfig_s {
-    adcChannelConfig_t vbat;
-    adcChannelConfig_t rssi;
-    adcChannelConfig_t current;
-    adcChannelConfig_t external1;
-    int8_t device; // ADCDevice
-} adcConfig_t;
 
 void adcInit(const adcConfig_t *config);
 uint16_t adcGetChannel(uint8_t channel);
