+++ src/main/drivers/adc_stm32f4xx.c
@@ -22,14 +22,15 @@
 #include "platform.h"
 
 #include "drivers/accgyro/accgyro.h"
+
+#include "drivers/dma.h"
+#include "drivers/io.h"
+#include "drivers/io_impl.h"
+#include "drivers/rcc.h"
+#include "drivers/sensor.h"
 #include "drivers/system.h"
 
-#include "drivers/io.h"
-#include "io_impl.h"
-#include "rcc.h"
-#include "dma.h"
-
-#include "drivers/sensor.h"
+#include "parameter_groups/adc.h"
 
 #include "adc.h"
 #include "adc_impl.h"
